+++ packages/babel-preset-sui/src/index.js
@@ -11,12 +11,7 @@ const plugins = (api, {useESModules = true} = {}) => [
   require('@babel/plugin-syntax-export-namespace-from').default,
   [require('@babel/plugin-proposal-decorators').default, {legacy: true}],
   [require('@babel/plugin-proposal-class-properties').default, {loose: true}],
-  [
-    require('babel-plugin-transform-react-remove-prop-types').default,
-    {
-      removeImport: true
-    }
-  ],
+  require('babel-plugin-transform-react-remove-prop-types').default,
   [
     require('@babel/plugin-transform-runtime').default,
     {