+++ packages/sui-bundler/loaders/linkLoaderConfigBuilder.js
@@ -1,9 +1,9 @@
 const fg = require('fast-glob')
 const path = require('path')
 
-const createSassLinkLoader = require('./sassLinkLoader')
 const log = require('../shared/log')
 const {defaultAlias} = require('../shared/resolve-alias')
+const createSassLinkImporter = require('./sassLinkImporter.js')
 
 const diccFromAbsolutePaths = (paths, init = {}) =>
   paths.reduce((acc, pkg) => {
@@ -42,36 +42,56 @@ module.exports = ({config, packagesToLink, linkAll}) => {
     diccFromAbsolutePaths(absolutePathForMonoRepo(linkAll))
   )
 
-  const sassLoaderWithLinkLoaderConfig = {
-    loader: require.resolve('sass-loader'),
-    options: {
-      sassOptions: {
-        importer: createSassLinkLoader(entryPoints)
-      }
-    }
-  }
-
-  const javascriptLinkLoader = {
-    test: /\.jsx?$/,
+  /**
+   * Create a loader config for javascript and css files that
+   * are handled by Webpack. So when we try to load them
+   * we check the entryPoints to change the source file
+   * if neccesary
+   */
+  const linkLoader = {
+    test: /\.(jsx?|scss)$/,
+    enforce: 'pre', // this will ensure is execute before transformations
     use: {
-      loader: 'link-loader',
+      loader: require.resolve('./LinkLoader'),
       options: {
         entryPoints
       }
     }
   }
 
-  const newRules = config.module.rules.map(rule => {
-    const {use: originalUseList} = rule
-    const use = originalUseList.map(singleUse => {
-      if (typeof singleUse === 'string' && singleUse.includes('sass-loader')) {
-        return sassLoaderWithLinkLoaderConfig
+  /**
+   * Create a sass-loader config for scss files that
+   * are handled by Sass. These are nested modules imported
+   * and thus is sass binary which needs a special config for them.
+   */
+  const sassLoaderWithLinkImporter = {
+    loader: require.resolve('sass-loader'),
+    options: {
+      sassOptions: {
+        importer: createSassLinkImporter(entryPoints)
       }
-      return singleUse
-    })
-    return {...rule, use}
+    }
+  }
+
+  /**
+   * Iterate over rules to change the previous sassLoader config
+   * with the one with the importer created
+   */
+  const {rules} = config.module
+  const rulesWithLink = rules.map(rule => {
+    const {use, test: regex} = rule
+    if (!regex.test('.css')) return rule
+
+    return {
+      ...rule,
+      use: [...use.slice(0, -1), sassLoaderWithLinkImporter]
+    }
   })
 
+  /**
+   * Return the new webpack config to be used
+   * with all the needed changes for linking packages
+   */
   return {
     ...config,
     resolve: {
@@ -84,16 +104,7 @@ module.exports = ({config, packagesToLink, linkAll}) => {
     },
     module: {
       ...config.module,
-      rules: [...newRules, javascriptLinkLoader]
-    },
-    resolveLoader: {
-      alias: {
-        ...config.resolveLoader.alias,
-        'link-loader': require.resolve('./LinkLoader'),
-        'externals-manifest-loader': require.resolve(
-          './ExternalsManifestLoader'
-        )
-      }
+      rules: [...rulesWithLink, linkLoader]
     }
   }
 }