+++ packages/sui-bundler/loaders/sassLinkImporter.js
@@ -1,4 +1,9 @@
-const createSassLinkLoader = entryPoints => url => {
+/**
+ * Create a Sass importer based on https://github.com/sass/node-sass#importer--v200---experimental
+ * @param {{ [packageName: string]: string}} entryPoints
+ * @return {(url: string) => {file: string} | null}
+ */
+const createSassLinkImporter = entryPoints => url => {
   if (Object.keys(entryPoints).find(pkg => url.match(`${pkg}/`))) {
     const [org, name] = url.split(/\//)
     const pkg = [org.replace('~', ''), name].join('/')
@@ -12,4 +17,4 @@ const createSassLinkLoader = entryPoints => url => {
   return null
 }
 
-module.exports = createSassLinkLoader
+module.exports = createSassLinkImporter