+++ packages/sui-ci/bin/sui-ci-release.js
@@ -12,12 +12,11 @@ const {
 // to keep compatibility with previous environment variable
 // GITHUB_TOKEN usage is preferred
 const gitHubToken = GITHUB_TOKEN || GH_TOKEN
+const isPullRequest = checkIsPullRequest()
 
-checkIsPullRequest().then(isPullRequest => {
-  release({
-    gitHubEmail,
-    gitHubToken,
-    gitHubUser,
-    isPullRequest
-  })
+release({
+  gitHubEmail,
+  gitHubToken,
+  gitHubUser,
+  isPullRequest
 })