+++ packages/sui-studio/src/app.js
@@ -1,7 +1,10 @@
 import {render} from 'react-dom'
 import {Router} from '@s-ui/react-router'
 
+import {importGlobals} from './components/tryRequire'
 import routes from './routes'
 import './index.scss'
 
-render(<Router routes={routes} />, document.getElementById('root'))
+importGlobals().then(() => {
+  render(<Router routes={routes} />, document.getElementById('root'))
+})